requirements.txt
jinja2
numpy
omegaconf
plotly_express
pandas
scipy
scikit-learn